lesson-10-evaluating-agents-with-python/exercises/starter/tools/pricing.py
# tools/pricing.py
from semantic_kernel.functions import kernel_function
import requests
import logging
from typing import Dict, Any
import json

logger = logging.getLogger(__name__)

class PricingTools:
    """Tools for dynamic pricing and competitor analysis using external APIs"""
    
    def __init__(self):
        # Using external pricing APIs for market analysis
        # For demo purposes, we'll simulate API calls to pricing services
        self.pricing_api_base = "https://api.pricingengine.com"
        self.competitor_api_base = "https://api.competitorpricing.com"
    
    @kernel_function(name="get_market_pricing", description="Get market pricing data using external pricing API")
    def get_market_pricing(self, product_name: str, category: str = None) -> Dict[str, Any]:
        """
        Get market pricing data for a product using external pricing API.
        
        Args:
            product_name: Name of the product to get pricing for
            category: Product category (optional)
            
        Returns:
            Dictionary containing market pricing information
        """
        try:
            logger.info(f"Getting market pricing via external API for: {product_name}")
            
            # Simulate API call to pricing service
            # In a real implementation, this would call actual pricing APIs
            pricing_api_response = {
                "status": "success",
                "timestamp": "2024-01-15T12:00:00Z",
                "product_name": product_name,
                "category": category or "Electronics",
                "market_data": {
                    "average_price": 89.99,
                    "price_range": {
                        "min": 49.99,
                        "max": 149.99
                    },
                    "competitor_prices": [
                        {
                            "retailer": "Amazon",
                            "price": 79.99,
                            "availability": "In Stock",
                            "rating": 4.5,
                            "url": "https://amazon.com/product"
                        },
                        {
                            "retailer": "Best Buy",
                            "price": 99.99,
                            "availability": "In Stock", 
                            "rating": 4.3,
                            "url": "https://bestbuy.com/product"
                        },
                        {
                            "retailer": "Walmart",
                            "price": 89.99,
                            "availability": "Limited Stock",
                            "rating": 4.2,
                            "url": "https://walmart.com/product"
                        }
                    ],
                    "price_trends": {
                        "last_30_days": "stable",
                        "last_90_days": "decreasing",
                        "trend_percentage": -5.2
                    },
                    "recommendations": {
                        "suggested_price": 84.99,
                        "reasoning": "Competitive pricing with 5% margin",
                        "confidence": 85
                    }
                }
            }
            
            return {
                "api_source": "Market Pricing API",
                "api_endpoint": f"{self.pricing_api_base}/v1/market-pricing",
                "pricing_analysis": pricing_api_response["market_data"]
            }
            
        except Exception as e:
            logger.error(f"❌ Failed to get market pricing via external API: {e}")
            return {
                "api_source": "Market Pricing API",
                "api_endpoint": f"{self.pricing_api_base}/v1/market-pricing",
                "pricing_analysis": {
                    "product_name": product_name,
                    "error": f"API call failed: {e}"
                }
            }
    
    @kernel_function(name="calculate_dynamic_price", description="Calculate dynamic pricing using external pricing engine")
    def calculate_dynamic_price(self, product_id: str, base_price: float, 
                               demand_factor: float = 1.0, inventory_level: int = 100) -> Dict[str, Any]:
        """
        Calculate dynamic pricing based on demand and inventory using external pricing engine.
        
        Args:
            product_id: Product ID to calculate pricing for
            base_price: Base price of the product
            demand_factor: Demand multiplier (1.0 = normal, >1.0 = high demand)
            inventory_level: Current inventory level
            
        Returns:
            Dictionary containing dynamic pricing calculation
        """
        try:
            logger.info(f"Calculating dynamic pricing via external API for product: {product_id}")
            
            # Simulate API call to dynamic pricing engine
            pricing_engine_response = {
                "status": "success",
                "timestamp": "2024-01-15T12:00:00Z",
                "product_id": product_id,
                "pricing_calculation": {
                    "base_price": base_price,
                    "demand_factor": demand_factor,
                    "inventory_level": inventory_level,
                    "calculated_price": round(base_price * demand_factor * (1 + (100 - inventory_level) / 1000), 2),
                    "price_adjustments": {
                        "demand_adjustment": round((demand_factor - 1.0) * base_price, 2),
                        "inventory_adjustment": round((100 - inventory_level) / 1000 * base_price, 2),
                        "total_adjustment": round((demand_factor - 1.0) * base_price + (100 - inventory_level) / 1000 * base_price, 2)
                    },
                    "pricing_strategy": "demand_and_inventory_based",
                    "confidence_score": 92,
                    "recommendations": [
                        "Price optimized for current market conditions",
                        "Consider promotional pricing if inventory exceeds 150 units",
                        "Monitor competitor pricing for validation"
                    ]
                }
            }
            
            return {
                "api_source": "Dynamic Pricing Engine API",
                "api_endpoint": f"{self.pricing_api_base}/v1/dynamic-pricing",
                "dynamic_pricing": pricing_engine_response["pricing_calculation"]
            }
            
        except Exception as e:
            logger.error(f"❌ Failed to calculate dynamic pricing via external API: {e}")
            return {
                "api_source": "Dynamic Pricing Engine API",
                "api_endpoint": f"{self.pricing_api_base}/v1/dynamic-pricing",
                "dynamic_pricing": {
                    "product_id": product_id,
                    "error": f"API call failed: {e}"
                }
            }
    
    @kernel_function(name="get_competitor_analysis", description="Get competitor pricing analysis using external API")
    def get_competitor_analysis(self, product_name: str, competitors: str = None) -> Dict[str, Any]:
        """
        Get competitor pricing analysis using external competitor analysis API.
        
        Args:
            product_name: Name of the product to analyze
            competitors: Comma-separated list of competitors to analyze
            
        Returns:
            Dictionary containing competitor analysis
        """
        try:
            logger.info(f"Getting competitor analysis via external API for: {product_name}")
            
            # Simulate API call to competitor analysis service
            competitor_api_response = {
                "status": "success",
                "timestamp": "2024-01-15T12:00:00Z",
                "product_name": product_name,
                "competitor_analysis": {
                    "total_competitors": 5,
                    "price_analysis": {
                        "our_price": 99.99,
                        "market_average": 89.99,
                        "price_position": "above_average",
                        "price_difference_percentage": 11.1
                    },
                    "competitor_breakdown": [
                        {
                            "competitor": "Amazon",
                            "price": 79.99,
                            "market_share": 35,
                            "price_advantage": "Lower by $20.00",
                            "strengths": ["Fast shipping", "Prime benefits"],
                            "weaknesses": ["Limited product support"]
                        },
                        {
                            "competitor": "Best Buy",
                            "price": 99.99,
                            "market_share": 25,
                            "price_advantage": "Same price",
                            "strengths": ["Expert advice", "In-store pickup"],
                            "weaknesses": ["Limited online selection"]
                        },
                        {
                            "competitor": "Walmart",
                            "price": 89.99,
                            "market_share": 20,
                            "price_advantage": "Lower by $10.00",
                            "strengths": ["Low prices", "Wide availability"],
                            "weaknesses": ["Limited product expertise"]
                        }
                    ],
                    "recommendations": [
                        "Consider price reduction to $89.99 to match Walmart",
                        "Emphasize product expertise and support as differentiators",
                        "Monitor Amazon's pricing for competitive response"
                    ],
                    "market_insights": {
                        "price_sensitivity": "medium",
                        "brand_loyalty": "high",
                        "feature_importance": "high"
                    }
                }
            }
            
            return {
                "api_source": "Competitor Analysis API",
                "api_endpoint": f"{self.competitor_api_base}/v1/analysis",
                "competitor_analysis": competitor_api_response["competitor_analysis"]
            }
            
        except Exception as e:
            logger.error(f"❌ Failed to get competitor analysis via external API: {e}")
            return {
                "api_source": "Competitor Analysis API",
                "api_endpoint": f"{self.competitor_api_base}/v1/analysis",
                "competitor_analysis": {
                    "product_name": product_name,
                    "error": f"API call failed: {e}"
                }
            }
